requirements.txt
google-auth==2.25.1
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.110.0
python-dotenv==1.0.0
